add debugging

+++ Kernel/include/processes.h
@@ -17,7 +17,7 @@ typedef enum
 
 typedef struct process_t
 {
-    uint32_t pid;
+    uint64_t pid;
     process_state_t state;
     uint32_t *stack;         // base stack pointer
     uint32_t *stack_pointer; // stack pointer
@@ -36,7 +36,7 @@ typedef process_t *process;
  * @param argv argumentos de entry_point
  * @return pid del proceso si es exitoso, 0 caso contrario
  */
-int createProcess(process memoryForProcess, char *process_name, int process_pid, int process_priority, void (*entry_point)(void), int argc, char *argv[]);
+int createProcess(process memoryForProcess, char *process_name, uint64_t process_pid, int process_priority, void (*entry_point)(void), int argc, char *argv[]);
 
 /**
  * Mata un proceso
