+++ Kernel/processes.c
@@ -28,7 +28,7 @@ typedef struct stack_frame
     uint64_t ss;
 } stack_frame;
 
-int createProcess(process memoryForProcess, char *process_name, int process_pid, int process_priority, void (*entry_point)(void), int argc, char *argv[]) // veamos si el entry_point puede recibir argumentos
+int createProcess(process memoryForProcess, char *process_name, uint64_t process_pid, int process_priority, void (*entry_point)(void), int argc, char *argv[]) // veamos si el entry_point puede recibir argumentos
 {
 
     memoryForProcess->pid = process_pid;
